+++ ChangeLog.md
@@ -9,6 +9,9 @@
 - Added the `Previous` command, which moves to the previous article
   regardless of read status.
   ([#12](https://github.com/davep/oldnews/pull/12))
+- Added the `MarkAllRead` command, which marks all unread articles in the
+  current category as read.
+  ([#13](https://github.com/davep/oldnews/pull/13))
 
 ## v0.0.2